+++ course_pages/locale/fr/LC_MESSAGES/djangojs.po
@@ -1,0 +1,54 @@
+
+msgid ""
+msgstr ""
+"Project-Id-Version: France Université Numérique 24\n"
+"Report-Msgid-Bugs-To: team@openfun.fr\n"
+"POT-Creation-Date: 2015-11-26 17:27+0100\n"
+"PO-Revision-Date: 2015-11-26 17:27+0000\n"
+"Last-Translator: FUN <team@openfun.fr>\n"
+"Language-Team: FUN <team@openfun.fr>\n"
+"Plural-Forms: nplurals=2; plural=(n > 1)\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=utf-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Generated-By: Babel 1.3\n"
+
+#: /edx/app/edxapp/fun-apps/course_pages/templates/courses_pages/js/small-course-block.underscore:33
+msgctxt "course-block"
+msgid "Ended"
+msgstr "Terminé"
+
+#: /edx/app/edxapp/fun-apps/course_pages/templates/courses_pages/js/small-course-block.underscore:49
+msgid "Learn more"
+msgstr "En savoir plus"
+
+#: /edx/app/edxapp/fun-apps/course_pages/templates/courses_pages/js/small-course-block.underscore:37
+msgctxt "course-block"
+msgid "Started"
+msgstr "En cours"
+
+#: /edx/app/edxapp/fun-apps/course_pages/templates/courses_pages/js/small-course-block.underscore:41
+msgctxt "course-block"
+msgid "Starts"
+msgstr "Débute"
+
+#: /edx/app/edxapp/fun-apps/course_pages/templates/courses_pages/js/filter.underscore:14
+msgid "Your search"
+msgstr "Votre recherche"
+
+#: /edx/app/edxapp/fun-apps/course_pages/templates/courses_pages/js/filter.underscore:53
+#: /edx/app/edxapp/fun-apps/course_pages/templates/courses_pages/js/filter.underscore:67
+msgid "all"
+msgstr "tous"
+
+#: /edx/app/edxapp/fun-apps/course_pages/templates/courses_pages/js/filter.underscore:55
+#: /edx/app/edxapp/fun-apps/course_pages/templates/courses_pages/js/filter.underscore:65
+msgid "per page"
+msgstr "par page"
+
+#: /edx/app/edxapp/fun-apps/course_pages/templates/courses_pages/js/filter.underscore:8
+msgid "result"
+msgid_plural "results"
+msgstr[0] "résultat"
+msgstr[1] "résultats"
+
