+++ course_pages/templates/course_pages/course-block/footer.html
