+++ course_pages/templates/course_pages/course-block/small.html
