+++ course_pages/templates/course_pages/js/filter.underscore
