+++ course_pages/templates/course_pages/js/small-course-block.underscore
