--- courses_api/locale/fr/LC_MESSAGES/django.po
@@ -1,106 +1,0 @@
-
-msgid ""
-msgstr ""
-"Project-Id-Version: France Université Numérique 38\n"
-"Report-Msgid-Bugs-To: team@openfun.fr\n"
-"POT-Creation-Date: 2015-11-18 11:31+0100\n"
-"PO-Revision-Date: 2015-11-18 11:31+0000\n"
-"Last-Translator: FUN <team@openfun.fr>\n"
-"Language-Team: FUN <team@openfun.fr>\n"
-"Plural-Forms: nplurals=2; plural=(n > 1)\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=utf-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Generated-By: Babel 1.3\n"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/index.html:52
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/index.html:592
-msgid "All themes"
-msgstr "Tous les thèmes"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/index.html:57
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/index.html:600
-msgid "All universities"
-msgstr "Tous les établissements"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/index.html:47
-msgid "Availability"
-msgstr "Disponibilité"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/index.html:63
-msgid "Certified training"
-msgstr "Formation certifiante"
-
-# Translated in edx by 'Cours'
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/index.html:11
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/index.html:540
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/index.html:546
-msgid "Courses"
-msgstr "Les cours"
-
-#: /edx/app/edxapp/fun-apps/courses_api/views.py:33
-#: /edx/app/edxapp/fun-apps/courses_api/tests/test_feed.py:35
-msgid "Fun latest published courses"
-msgstr "Derniers cours publiés sur FUN"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/index.html:48
-msgid "Imminent end of enrollment"
-msgstr "Fin d'inscription imminente"
-
-#: /edx/app/edxapp/fun-apps/courses_api/views.py:35
-msgid "Latests courses published on www.france-universite-numerique-mooc.fr"
-msgstr "Derniers cours publiés sur www.france-universite-numerique-mooc.fr"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/course-block/big.html:45
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/course-block/small.html:45
-msgid "Learn more"
-msgstr "En savoir plus"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/index.html:50
-msgid "New"
-msgstr "Nouveau"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/feed/feed.html:10
-msgid "Proposed by"
-msgstr "Cours proposé par"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/index.html:559
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/index.html:569
-msgid "Refine search results"
-msgstr "Affiner les résultats"
-
-# Translated in edx by 'Commence'
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/course-block/big.html:36
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/course-block/small.html:36
-msgctxt "course-block"
-msgid "Starts"
-msgstr "Débute"
-
-msgctxt "course-block"
-msgid "Started"
-msgstr "En cours"
-
-msgctxt "course-block"
-msgid "Ended"
-msgstr "Terminé"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/index.html:49
-msgid "Starts soon"
-msgstr "Démarre bientôt"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/index.html:52
-msgid "Themes"
-msgstr "Thèmes"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/index.html:57
-msgid "Universities"
-msgstr "Etablissements"
-
-# Translated in edx by 'Université'
-#: /edx/app/edxapp/fun-apps/courses/forms.py:29
-msgid "University"
-msgstr "Établissement"
-
-#~ msgid "Start date:"
-#~ msgstr "Date de début :"
-
