--- courses_api/locale/fr/LC_MESSAGES/djangojs.po
@@ -1,52 +1,0 @@
-
-msgid ""
-msgstr ""
-"Project-Id-Version: France Université Numérique 21\n"
-"Report-Msgid-Bugs-To: team@openfun.fr\n"
-"POT-Creation-Date: 2015-11-18 11:31+0100\n"
-"PO-Revision-Date: 2015-11-18 11:31+0000\n"
-"Last-Translator: FUN <team@openfun.fr>\n"
-"Language-Team: FUN <team@openfun.fr>\n"
-"Plural-Forms: nplurals=2; plural=(n > 1)\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=utf-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Generated-By: Babel 1.3\n"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/js/small-course-block.underscore:49
-msgid "Learn more"
-msgstr "En savoir plus"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/js/small-course-block.underscore:33
-msgctxt "course-block"
-msgid "Started"
-msgstr "En cours"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/js/small-course-block.underscore:41
-msgctxt "course-block"
-msgid "Starts"
-msgstr "Débute"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/js/small-course-block.underscore:37
-msgctxt "course-block"
-msgid "Ended"
-msgstr "Terminé"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/js/filter.underscore:31
-msgid "Your search"
-msgstr "Votre recherche"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/js/filter.underscore:46
-msgid "all"
-msgstr "tous"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/js/filter.underscore:44
-msgid "per page"
-msgstr "par page"
-
-#: /edx/app/edxapp/fun-apps/courses_api/templates/courses_api/js/filter.underscore:5
-msgid "result"
-msgid_plural "results"
-msgstr[0] "résultat"
-msgstr[1] "résultats"
-
