+++ courses_api/routers.py
@@ -3,7 +3,7 @@ from rest_framework.routers import DefaultRouter
 
 class CourseAPIRouter(DefaultRouter):
 
-    BASENAME = 'api'
+    BASENAME = 'courses'
 
     def get_default_base_name(self, viewset):
         return self.BASENAME
