+++ fun/lms/urls.py
@@ -21,8 +21,11 @@ urlpatterns = patterns('',
     (r'^help/', include('faq.urls', namespace='faq')),
     (r'^backoffice/', include('backoffice.urls', namespace='backoffice')),
 
+    # fun api
+    (r'^fun/', include('courses_api.urls', namespace='fun-courses-api')),
+
     # override edX's courses page to replace by FUN's one (we need to use a other route)
-    (r'^cours/', include('courses_api.urls', namespace='fun-courses')),
+    (r'^cours/', include('course_pages.urls', namespace='fun-courses')),
     url(  # intercept old routes, and redirect
         r'^courses/$',
         RedirectView.as_view(url=reverse_lazy('fun-courses:index'))
