+++ funsite/static/funsite/css/fun.css
@@ -1131,7 +1131,7 @@ ul.universities-list {
 
 
 
-/* Courses filtering and searching courses_api/index.html */
+/* Courses filtering and searching course_pages/index.html */
 
 .course-list .list ul {
     list-style-type: none;
